print_reg: Separate printing from extraction of information into 'stash'
This fixes a bug that makes 'cpuid -r' not print the 0x40000001 leaf,
because stash->hypervisor is never set when in raw mode.

This patch separates the code that changes fields in the 'stash' struct
from the code that prints that information. This way, the stash struct
will get updated even when in raw mode, so other parts of the code can
use that information.

Signed-off-by: Eduardo Habkost <ehabkost@raisama.net>